requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
configparser==5.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.1
django-allauth==0.50.0
django-appconf==1.0.5
django-cors-headers==3.13.0
django-filter==22.1
django-imagekit==4.1.0
django-js-asset==2.0.0
django-minio-storage==0.3.10
django-mptt==0.13.4
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.3
environs==9.5.0
future==0.18.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.17.1
mccabe==0.7.0
minio==6.0.2
oauthlib==3.2.0
packaging==21.3
pilkit==2.0
Pillow==9.2.0
psycopg2==2.9.3
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
